backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==14.1
openai==1.101.0
python-multipart==0.0.19
pydantic==2.10.4
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
haystack-ai==2.17.1
motor==3.6.0
pymongo==4.10.1
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
pydantic[email]